requirements.txt
git+https://github.com/Rapptz/discord.py
git+https://github.com/Gorialis/jishaku
mystbin.py
thefuzz[speedup]